Small update
Had some weirdness going on with a 'db.lock' file, so I got rid of the folder that it was related to, and my script seems to have come unattached? Not really good practice to randomly delete stuff but I had no idea where it came from and it wouldn't push to github with those files, nor could github revert the changes. Very strange, but deleting them seems to have worked?