Matching Part 1
Can match numbers up to 13 and use the King as a solo match. Next things to work on are making it so a failed match resets the cards in the match method, and then making it so you can match with the draw pile. Then I'll have to make it so you can only match with the draw pile and bottom row.