Matching Part 2
Made it so a failed match resets the cards in the match method, and then made it so you can match with the draw pile. Next I'll have to make the waste pile work and then make it so you can only match with the draw pile and bottom row.